requirements.txt
python-dotenv==1.0.1
aiohttp==3.9.5
aiogram==3.10.0
aiosqlite==0.20.0
APScheduler==3.10.4
pydub==0.25.1
